Add files via upload
Extensions of schelling model, extension models gentrification. Turns a number of blue turtles into "bobo's" (Bohemian Bourgeoisie). Bobo's are blue, but act like if they are orange. This means they want to be close to orange, but orange does not want to be close to them. Other blue wants to be close to them, but they do not want to be close to them. Bobo's are insects in the visualization. Results do not change much from base model.

Things to add:
- some kind of price on patches
- some kind of financial info for turtles
 - Orange and blue have different finances
- Price changes on patches where bobo's are